mymarket/requirements.txt
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
class-registry==2.1.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.57.0
django-cors-headers==4.3.0
django-filter==23.3
django-js-asset==2.1.0
django-mptt==0.15.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
duplicity==0.8.21
fasteners==0.14.1
filters==1.3.2
future==0.18.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==23.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.1
regex==2023.10.3
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
